requirements.txt
aiolimiter==1.0.0
anyio==3.6.2
APScheduler==3.10.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
docopt==0.6.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pipreqs==0.4.13
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarg==0.1.9